+++ src/SimpleSmtp_SSL_QT5/smtp_attachements/smtp.h
@@ -36,7 +36,7 @@ class Smtp : public QObject {
 
 public:
   Smtp(const QString &user, const QString &pass, const QString &host,
-       int port = 465, int timeout = 30000);
+       quint16 port = 465, int timeout = 30000);
   ~Smtp();
 
   void sendMail(const QString &from, const QString &to, const QString &subject,
@@ -61,7 +61,7 @@ private:
   QString pass;
   QString host;
   static QString status;
-  int port;
+  quint16 port;
   enum states {
     Tls,
     HandShake,
